Make multiqc/__main__.py a script. Import run() in __init__.py. Remove multiqc/multiqc.py
